Add encoding to ReadAllText call
When reading the contents of a non-unicode file, e.g. code page 1252,
non-ascii text is sometimes replaced by a placeholder character such as ?.

In order to fix this we add the specified encoding to the call to
ReadAllText. This way we both read and write (using posh command Out-File)
using the same encoding.